% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{calcOddsRatio}
\alias{calcOddsRatio}
\title{calcOddsRatio}
\usage{
calcOddsRatio(table, alpha = 0.05, referencerow = 2, quiet = FALSE)
}
\arguments{
\item{table}{Contingency table.}

\item{alpha}{critical value}

\item{referencerow}{fix a indep. group as a base/reference.}

\item{quiet}{If there are just two treatments (exposed/nonexposed). Default value is FALSE.}
}
\value{

}
\description{
Calculates odds Ratio and CI.
}
