% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{calcRelativeRisk}
\alias{calcRelativeRisk}
\title{calcRelativeRisk}
\usage{
calcRelativeRisk(table, alpha = 0.05, referencerow = 2, roundoff = 3)
}
\arguments{
\item{table}{Contingency table.}

\item{alpha}{critical value}

\item{referencerow}{fix a indep. group as a base/reference.}

\item{roundoff}{round-off digit.}
}
\value{
character
}
\description{
Calculates the Relative Risk and CI.
}
