% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{freq_table}
\alias{freq_table}
\title{freq_table}
\usage{
freq_table(x, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{input numeric variable}

\item{xname}{name of the variable}
}
\value{
a data frame of the frequency table
}
\description{
It gives a frequency table with the percentage.
}
