% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mycat}
\alias{mycat}
\title{mycat}
\usage{
mycat(text)
}
\arguments{
\item{text}{Table's heading}
}
\value{

}
\description{
It gives the designed table heading.
}
\examples{
mycat("This is my table")
}
