% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mycenter}
\alias{mycenter}
\title{mycenter}
\usage{
mycenter(text, lsp, usp)
}
\arguments{
\item{text}{Table's heading}

\item{lsp}{no. of spaces/blanks spaces in the left side of the text.}

\item{usp}{no. of spaces/blanks spaces in the right side of the text.}
}
\value{
character
}
\description{
it helps to fix the table's heading into the center.
}
\examples{
mycat(mycenter("This is my Table"),20,0)
}
