% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mydescriptive}
\alias{mydescriptive}
\title{mydescriptive}
\usage{
mydescriptive(x)
}
\arguments{
\item{x}{input numeric variable}
}
\value{
character vector
}
\description{
It gives the general summary statistics:  no. of observations, average, standard deviation, median and interquartile range with 1st quartile and 3rd quartile.
}
\examples{
x=c(1:10)
  mydescriptive(x)
}
