% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mytab_indep_two_Populations}
\alias{mytab_indep_two_Populations}
\title{mytab_indep_two_Populations}
\usage{
mytab_indep_two_Populations(x, group, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{input dependent numeric variable.}

\item{group}{independent character vector for the names of the two populations.}

\item{xname}{name of the input dependent numeric variable.}
}
\value{
Returns a data frame containing the descriptive stat along with the indep. t-test stat., indep. t-test P Value & Non parametric(Mann Whitney U test) P Value.
}
\description{
It gives a output table for testing the significant difference between the  means of two independent populations.
}
