% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mytab_indp_K_Populations}
\alias{mytab_indp_K_Populations}
\title{mytab_indp_K_Populations}
\usage{
mytab_indp_K_Populations(x, group, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{dependent numeric variable}

\item{group}{independent categorical variable}

\item{xname}{name of the dependent variable}
}
\value{
Returns a data frame containing the descriptive stat along with the 1 way ANOVA F test stat., 1 way ANOVA P Value & Non parametric(Kruskal Wallis test) P Value.
}
\description{
It gives a output table for testing the significant difference between the  means of k number of independent populations.
}
