% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/My_functions.R
\name{mytab_paired_one_Populations}
\alias{mytab_paired_one_Populations}
\title{mytab_paired_one_Populations}
\usage{
mytab_paired_one_Populations(x.before, x.after, var)
}
\arguments{
\item{x.before}{numeric vector for the pre score.}

\item{x.after}{numeric vector for the post score.}

\item{var}{name of the variable/ population.}
}
\value{
Returns a data frame containing the descriptive stat along with the Paired.t- test test statistic, Confidence Interval, Paired.t- test P-value & Non parametric(Wilcoxon Signed Rank test) P Value.
}
\description{
It gives a output table for testing the significant difference between the pre and post means of a independent populations.
}
